refactor: Move root widget to lib/src/app.dart
